requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP (Chatbot)
nltk>=3.8.0

# Deep Learning (Vision pipeline)
tensorflow>=2.15.0; platform_machine != "arm64" or platform_system != "Darwin"
tensorflow-macos>=2.15.0; platform_machine == "arm64" and platform_system == "Darwin"
tensorflow-metal>=1.1.0; platform_machine == "arm64" and platform_system == "Darwin"

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data acquisition
kaggle>=1.5.0

# Jupyter ecosystem
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# PDF export (optional)
reportlab>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
